# backend/api/rest/v1/swagger.yaml
definitions:
...
  github_com_JMURv_sso_internal_dto.CreatePermissionRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  github_com_JMURv_sso_internal_dto.CreateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permissions:
        items:
          type: integer
        type: array
    required:
    - name
    type: object
...
  github_com_JMURv_sso_internal_dto.PaginatedPermissionResponse:
    properties:
      count:
        type: integer
      current_page:
        type: integer
      data:
        items:
          $ref: '#/definitions/github_com_JMURv_sso_internal_models.Permission'
        type: array
      has_next_page:
        type: boolean
      total_pages:
        type: integer
    type: object
  github_com_JMURv_sso_internal_dto.PaginatedRoleResponse:
    properties:
      count:
        type: integer
      current_page:
        type: integer
      data:
        items:
          $ref: '#/definitions/github_com_JMURv_sso_internal_models.Role'
        type: array
      has_next_page:
        type: boolean
      total_pages:
        type: integer
    type: object
...
  github_com_JMURv_sso_internal_dto.UpdatePermissionRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  github_com_JMURv_sso_internal_dto.UpdateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permissions:
        items:
          type: integer
        type: array
    required:
    - name
    type: object
  github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse:
    properties:
      errors:
        items:
          type: string
        type: array
    type: object
...
  github_com_JMURv_sso_internal_models.Permission:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_JMURv_sso_internal_models.Role:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_JMURv_sso_internal_models.Permission'
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /perm:
    get:
      description: Retrieve a paginated list of permissions with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_dto.PaginatedPermissionResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: List permissions
      tags:
      - Permission
    post:
      consumes:
      - application/json
      description: Add a new permission to the system
      parameters:
      - description: Permission details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permission ID
          schema:
            type: int
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "409":
          description: permission already exists
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Create a new permission
      tags:
      - Permission
  /perm/{id}:
    delete:
      description: Remove a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: invalid ID
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: permission not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Delete a permission
      tags:
      - Permission
    get:
      description: Retrieve a specific permission
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_models.Permission'
        "400":
          description: invalid ID
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: permission not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Get permission by ID
      tags:
      - Permission
    put:
      consumes:
      - application/json
      description: Modify an existing permission
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated permission data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.UpdatePermissionRequest'
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: permission not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Update a permission
      tags:
      - Permission
  /roles:
    get:
      description: Retrieve a paginated list of roles with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_dto.PaginatedRoleResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: List roles
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Add a new role with associated permissions
      parameters:
      - description: Role details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.CreateRoleRequest'
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Role ID
          schema:
            type: int
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "409":
          description: role already exists
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Create a new role
      tags:
      - Role
  /roles/{id}:
    delete:
      description: Remove a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: invalid ID
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: role not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Delete a role
      tags:
      - Role
    get:
      description: Retrieve details of a specific role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_models.Role'
        "400":
          description: invalid ID
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: role not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Get role by ID
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Modify an existing role's name or permissions
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated role data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.UpdateRoleRequest'
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: role not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Update a role
      tags:
      - Role
...
swagger: "2.0"